requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Flask-SocketIO==5.3.4
eventlet==0.33.3
Pillow==10.0.0
bcrypt==4.0.1
pymysql==1.1.0
eventlet==0.33.3